remove some compilation warnings

+++ test/Tdna4_revcomp.cxx
@@ -35,7 +35,7 @@ TEST_CASE("Basic revcomp checks")
 	memset(buffer, 0, forward.size() + 1);
 
 	auto end_ptr = dna4_revcomp(dna::begin(forward), dna::end(forward), buffer);
-	REQUIRE(end_ptr - buffer == forward.size());
+	REQUIRE(static_cast<size_t>(end_ptr - buffer) == forward.size());
 	REQUIRE(forward == buffer);
 
 	delete[] buffer;
